refactor(node): extract mempool subcommands to commands/mempool.rs
Two cmd_mempool_* functions move out of main.rs:
  cmd_mempool_clear, cmd_mempool_stats

mempool clear is destructive (drops every pending tx) but local-only;
peers re-broadcast their own copies on the next tick. Module docstring
calls this out so the next person doesn't have to guess.
